Add client authentication configuration.
Adds the ability to connect to a secured NiFi or NiFi Registry instance
using two methods:

- Mutual authentication via two-way TLS (can now configure SSLContext)
- Username/password (e.g., LDAP) authentication

A number of bugs in the swager-generated client, particularly for the
registry client, have also been adressed.

In order to make these improvements, modifications were made to the
stock swagger-codegen Python generator tempalates that were originally
used to bootstrap the REST API clients. The modified templates, along
with instructions for how to used them to generate new client code as
the Swagger specs evolve, were added to this repository. A note of
attribution to the swagger-codegen project was added to the LICENSE.