Issue #100: create service users and groups (#102)
* Modify create_service_user to pass a revision when the service is nifi

Creating a new user requires a revision with version = 0.

* Create new create_service_user_group function

Create a new function for creating user groups, optionally populated with users

* Change user default from empty list to None

* Modify lines to be less than 80 characters long

* Add tests for create_service_user and create_service_user_group for nifi and registry

* Disable tests for user/group creation due to secure environment requirement

Creating users and groups requires a secure instance of NiFi.  Current testing does not include such an instance.  These tests can be re-enabled if the test environment provides a secure instance in the future.